requirements.txt
mysql-connector-python
Flask
flask_cors